refactor database pkg & store dbpool in env struct
