+++ server/database/go.mod
@@ -1,8 +1,11 @@
-module github.com/TingeOGinge/INSE-2C
+module github.com/TingeOGinge/INSE-2C/server/database
 
 go 1.17
 
-require github.com/TingeOGinge/INSE-2C/server/database v0.0.0-00010101000000-000000000000
+require (
+	github.com/jackc/pgx v3.6.2+incompatible
+	github.com/jackc/pgx/v4 v4.13.0
+)
 
 require (
 	github.com/jackc/chunkreader/v2 v2.0.1 // indirect
@@ -12,12 +15,8 @@ require (
 	github.com/jackc/pgproto3/v2 v2.1.1 // indirect
 	github.com/jackc/pgservicefile v0.0.0-20200714003250-2b9c44734f2b // indirect
 	github.com/jackc/pgtype v1.8.1 // indirect
-	github.com/jackc/pgx v3.6.2+incompatible // indirect
-	github.com/jackc/pgx/v4 v4.13.0 // indirect
 	github.com/jackc/puddle v1.1.3 // indirect
 	github.com/pkg/errors v0.9.1 // indirect
 	golang.org/x/crypto v0.0.0-20210711020723-a769d52b0f97 // indirect
 	golang.org/x/text v0.3.6 // indirect
 )
-
-replace github.com/TingeOGinge/INSE-2C/server/database => ./database
