+++ server/ms_entry.go
@@ -1,14 +1,24 @@
 package main
 
 import (
-	"net/http"
-	"log"
-	"fmt"
+	"context"
 	"encoding/json"
+	"fmt"
+	"log"
+	"net/http"
+
+	"github.com/TingeOGinge/INSE-2C/server/database"
+	"github.com/jackc/pgx/v4/pgxpool"
 )
 
-func search (w http.ResponseWriter, r *http.Request) {
-	recipes, err := Query("mainSearch", r.URL.Query()["searchTerms"])
+type Env struct {
+	ecochef interface {
+		MainSearch(parameters []string) ([]database.Recipe, error)
+	}
+}
+
+func (env *Env) searchHandler (w http.ResponseWriter, r *http.Request) {
+	recipes, err := env.ecochef.MainSearch(r.URL.Query()["searchTerms"])
 	if err != nil {
 		errString := fmt.Sprintf("Error executing query: %v\n", err)
 		http.Error(w, errString, http.StatusInternalServerError)
@@ -24,7 +34,6 @@ func search (w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	w.Header().Set("Content-Type", "application/json")
 	if _, err := w.Write(payload); err != nil {
 		errString := fmt.Sprintf("Error writing payload response: %v", err)
 		http.Error(w, errString, http.StatusInternalServerError)
@@ -33,8 +42,25 @@ func search (w http.ResponseWriter, r *http.Request) {
 	}
 }
 
-func App() {
-	http.HandleFunc("/api/mainSearch", search)
+func makeHandler(fn func(http.ResponseWriter, *http.Request)) http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		w.Header().Set("Content-Type", "application/json")
+		fn(w, r)
+	}
+}
+
+func main() {
+	url := "postgres://myuser:inse2c@localhost:5432/ecochefdb"
+	dbpool, err := pgxpool.Connect(context.Background(), url)
+	if err != nil {
+		log.Fatal(fmt.Sprintf("Unable to connect to database: %v\n", err))
+	}
+
+	env := &Env{
+		ecochef: database.EcochefEnv{DBpool: dbpool},
+	}
+
+	http.HandleFunc("/api/mainSearch", makeHandler(env.searchHandler))
 	http.Handle("/", http.FileServer(http.Dir("./front-end")))
 	
 	log.Println("Server started on port 5000")
